+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## [0.0.1] - 2020-10-8
 
 * Blur is a wrapper widget that blur it's child. There are ImageBlur.asset and ImageBlur.network that blur the image.
+
+## [1.0.0] - 02-11-2020
+
+* Breaking: The `borderRadius` datatype has changed from `double` to `BorderRadius`.
+* Add: `height` and `width` properties to `ImageBlur.asset` and `ImageBlur.network`.
+* Add: `Frost` class which blurs the background of the child.
+* Breaking: The `overlayChild` is renamed to `overlay`
+* Add: `alignment` property for the overlay child.